--- lib/control/constants.py
@@ -1,15 +1,0 @@
-# Physical constants.
-G = 32.17405                        # feet / second^2
-
-# Launch vehicle constants.
-VEHICLE_MASS = 1.1902231            # slugs
-
-# State determination constants.
-LAUNCH_ALTITUDE = 100               # feet
-LAUNCH_ACCELERATION = 200           # feet / second^2
-LAUNCH_ALTITUDE_CRITICAL = 300      # feet / second^2
-BURNOUT_ALTITUDE = 1000             # feet
-BURNOUT_ACCELERATION = 0            # feet / second^2
-BURNOUT_ALTITUDE_CRITICAL = 1500    # feet
-APOGEE_ALTITUDE = 5200              # feet
-APOGEE_VELOCITY = 0                 # feet
