+++ lib/physical/sensors.py
@@ -2,36 +2,15 @@ import time
 
 import adafruit_bmp3xx
 import adafruit_bno055
-import board
-
-from units import celsius_to_fahrenheit, meters_to_feet
-
-I2C = board.I2C()
-
-def sensor_reading(func):
-    # @wraps(func)
-    # def wrapper(self, *args, **kwargs):
-    #     reading = func(self, *args, **kwargs)
-    #     self.readings[func.__name__].append((time.time(), reading))
-    #     return reading
-    # return wrapper
-    return func
-        
 
 class _BMP390:
-    def __init__(self):
-        self.altimeter = adafruit_bmp3xx.BMP3XX_I2C(I2C)
+    def __init__(self, i2c):
+        self.altimeter = adafruit_bmp3xx.BMP3XX_I2C(i2c)
         self.altimeter.pressure_oversampling = 1
-
-    @sensor_reading
+    
     def temperature(self):
         return celsius_to_fahrenheit(self.altimeter.temperature)
-
-    @sensor_reading
-    def pressure(self):
-        return self.altimeter.pressure
-
-    @sensor_reading
+    
     def altitude(self):
         return meters_to_feet(self.altimeter.altitude)
 
@@ -47,8 +26,8 @@ class _BMP390:
 
 
 class _BNO055:
-    def __init__(self) -> None:
-        self.imu = adafruit_bno055.BNO055_I2C(I2C)
+    def __init__(self, i2c) -> None:
+        self.imu = adafruit_bno055.BNO055_I2C(i2c)
 
         # Configure the BNO055 to operate in 16g mode.
         # Unfortunately, a typo in the datasheet implied that
@@ -59,22 +38,24 @@ class _BNO055:
         self.imu.accel_range = adafruit_bno055.ACCEL_16G
         self.imu.mode = adafruit_bno055.AMG_MODE
 
-    @sensor_reading
     def acceleration(self):
         return tuple(map(meters_to_feet, self.imu.acceleration))
 
-    @sensor_reading
     def gyro(self):
         return tuple(map(meters_to_feet, self.imu.gyro))
         
-    @sensor_reading
     def magnetic(self):
         return tuple(map(meters_to_feet, self.imu.magnetic))
 
-    @sensor_reading
     def euler(self):
         return self.imu.euler
 
 
-ALTIMETER = _BMP390()
-IMU = _BNO055()
+def meters_to_feet(n):
+    return n * 3.28084
+
+def celsius_to_fahrenheit(n):
+    return n * 1.8 + 32
+
+def newtons_to_pounds(n):
+    return n * 0.224809
